rename AppModel to AppSettings
