+++ app/src/main/java/com/example/tonetuner_v2/audio/audioSources/MicSource.kt
@@ -2,13 +2,12 @@ package com.example.tonetuner_v2.audio.audioSources
 
 import android.Manifest
 import android.annotation.SuppressLint
-import android.content.Context
 import android.media.AudioFormat
 import android.media.AudioRecord
 import android.media.MediaRecorder
 import androidx.annotation.RequiresPermission
-import com.example.tonetuner_v2.app.AppModel.CAPTURE_BUFFER_SIZE
-import com.example.tonetuner_v2.app.AppModel.SAMPLE_RATE
+import com.example.tonetuner_v2.app.AppSettings.CAPTURE_BUFFER_SIZE
+import com.example.tonetuner_v2.app.AppSettings.SAMPLE_RATE
 import com.example.tonetuner_v2.extensions.hasPermission
 import com.example.tonetuner_v2.util.ContextHolder
 import com.example.tonetuner_v2.util.Logger
