+++ app/src/main/java/com/example/tonetuner_v2/audio/audioSources/SignalSource.kt
@@ -5,7 +5,7 @@ import com.example.signallib.SignalManager
 import com.example.signallib.SignalSettings
 import com.example.signallib.enums.Note
 import com.example.signallib.enums.WaveShape
-import com.example.tonetuner_v2.app.AppModel
+import com.example.tonetuner_v2.app.AppSettings
 import kotlin.random.Random
 
 class SignalSource(numHarmonics: Int) : AudioSource {
@@ -16,12 +16,12 @@ class SignalSource(numHarmonics: Int) : AudioSource {
     val signalSettings = SignalSettings(
         harmonicSeries = HarmonicSeries(numHarmonics),
         waveShape = WaveShape.SINE,
-        sampleRate = AppModel.SAMPLE_RATE,
-        bufferSize = AppModel.CAPTURE_BUFFER_SIZE
+        sampleRate = AppSettings.SAMPLE_RATE,
+        bufferSize = AppSettings.CAPTURE_BUFFER_SIZE
     )
     private val signalManager = SignalManager(signalSettings)
 
-    private var audioBuffer = FloatArray(AppModel.CAPTURE_BUFFER_SIZE)
+    private var audioBuffer = FloatArray(AppSettings.CAPTURE_BUFFER_SIZE)
 
     override fun getAudio(bufferSize: Int): List<Float> {
         if (bufferSize != audioBuffer.size)
@@ -37,7 +37,7 @@ class SignalSource(numHarmonics: Int) : AudioSource {
     }
 
     fun generateRandom() {
-        this.notes = setOf(AppModel.NOTE_RANGE.random())
+        this.notes = setOf(AppSettings.NOTE_RANGE.random())
         this.pitchBend = Random.nextDouble(-0.5, 0.5).toFloat()
         this.signalSettings.harmonicSeries.generateRandom()
     }
