+++ app/src/main/java/com/example/tonetuner_v2/ui/composables/FreezeButton.kt
@@ -11,7 +11,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.style.TextAlign
-import com.example.tonetuner_v2.app.AppModel
 
 @Composable
 fun FreezeButton(
